TP-4/EXERCICE-4.2-A-RENDRE/requirements.txt
streamlit
pillow
piexif
folium
streamlit-folium
exif